fix ctrl+backspace in options menu
- seems like a logic error from 70f1181edb4d4ab4a39e1bae511ad97d730e3de8
- also use char16_t consistently for search strings
- and some other very minor cleanup that shouldn't cause regressions (famous last words)

+++ src/Engine/Input/Mouse.cpp
@@ -166,7 +166,7 @@ void Mouse::onButtonChange(ButtonIndex button, bool down) {
     using enum ButtonIndex;
     if(button == BUTTON_NONE || button >= BUTTON_COUNT) return;
 
-    this->bMouseButtonDownArray[static_cast<size_t>(button)] = down;
+    this->bMouseButtonDownArray[button] = down;
 
     // notify listeners
     for(auto &listener : this->listeners) {
